requirements.txt
# Core Dependencies
streamlit==1.31.0
python-dotenv==1.0.0
requests==2.31.0

# LLM & AI
google-generativeai==0.3.2
langchain==0.1.9
langchain-community==0.0.24
langchain-google-genai==0.0.9

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Search & Scraping
google-search-results==2.4.2
duckduckgo-search==4.1.1
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# API & Web
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# Utilities
rich==13.7.0
colorama==0.4.6